backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[all]>=0.133.0",
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "prometheus-client>=0.24.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
